+++ tests/api/test_guids.py
@@ -2,7 +2,7 @@ import json
 import pytest
 
 from tests.api.constants import HTTP_STATUS_CODES_FOR_PUT, TEST_DATA_PATH
-from tests.api.request_executors.request_executor_fixture import request_executor
+from tests_lib.helpers.api.request_executors.request_executor_fixture import request_executor
 from tests_lib.common.custom_logger import CustomLogger
 from tests_lib.common.json_loader import load_json
 
@@ -59,13 +59,8 @@ def test_put_guids(status_code: int | str, logger_fixture: CustomLogger, request
     """
     logger_fixture.info(f"Starting test_put_get_guids with status_code: {status_code}")
     try:
-        # with open(TEST_DATA_PATH + "PUT_guids_positive.json", 'r') as expected_data_file:
-        #     expected_data_json = json.load(expected_data_file)
-
         expected_data_json = load_json(TEST_DATA_PATH + "PUT_guids_positive.json", logger_fixture)
         expected_data_json["status_code"] = status_code
-
-
         put_response = request_executor.execute_put(f"/guids", expected_data_json)
         put_response_json = put_response.json()
         assert put_response.status_code == 200
@@ -111,7 +106,6 @@ def test_put_guids_without_request_body(logger_fixture: CustomLogger, request_ex
     except AssertionError as e:
         logger_fixture.error(f"Test failed with error: {e}")
         raise
-
 
 
 def test_get_guids(logger_fixture: CustomLogger, request_executor):
